requirements.txt
streamlit
# streamlit_chat
# openai
google-generativeai